move demo code and switch to babel
