+++ .gitignore
@@ -6,6 +6,7 @@
 package-lock.json
 
 build/
+dist/
 node_modules/
 .DS_Store
 npm-debug.log*
