+++ demo/components/Square.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import withEventHandlers from '../lib/hoc/withEventHandlers';
+import withEventHandlers from '../../src/hoc/withEventHandlers';
 
 const StyledSquare = styled.div`
   top: ${props => (props.y)}px;
