+++ demo/components/Squares.jsx
