+++ package.json
@@ -4,10 +4,8 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "build": "neutrino build",
-    "start": "neutrino start",
-    "lint": "neutrino lint",
-    "test": "neutrino test"
+    "build:cjs": "mkdirp dist/cjs && babel src/ -d dist/cjs --copy-files --ignore '*.test.js'",
+    "build:esm": "mkdirp dist/esm && babel src/ -d dist/esm --copy-files --no-babelrc --presets=react --plugins=transform-object-rest-spread --ignore '*.test.js'"
   },
   "dependencies": {
     "prop-types": "^15.6.2",
@@ -17,9 +15,13 @@
     "styled-components": "^3.3.3"
   },
   "devDependencies": {
-    "@neutrinojs/airbnb": "^8.2.3",
-    "@neutrinojs/jest": "^8.2.3",
-    "@neutrinojs/react": "^8.2.3",
-    "neutrino": "^8.2.3"
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-plugin-styled-components": "^1.5.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1"
   }
 }
