+++ src/event-registry.js
